+++ ui/src/routes/devices.$id.settings.access._index.tsx
@@ -9,6 +9,7 @@ import { GridCard } from "@/components/Card";
 import { Button, LinkButton } from "@/components/Button";
 import { InputFieldWithLabel } from "@/components/InputField";
 import { SelectMenuBasic } from "@/components/SelectMenuBasic";
+import { SettingsItem } from "@components/SettingsItem";
 import { SettingsSectionHeader } from "@/components/SettingsSectionHeader";
 import { useDeviceUiNavigation } from "@/hooks/useAppNavigation";
 import notifications from "@/notifications";
@@ -18,7 +19,6 @@ import { isOnDevice } from "@/main";
 import { TextAreaWithLabel } from "@components/TextArea";
 
 import { LocalDevice } from "./devices.$id";
-import { SettingsItem } from "./devices.$id.settings";
 import { CloudState } from "./adopt";
 
 export interface TLSState {
