+++ ui/src/routes/devices.$id.settings.appearance.tsx
@@ -1,9 +1,9 @@
 import { useCallback, useState } from "react";
 
+import { SettingsItem } from "@components/SettingsItem";
+
 import { SettingsPageHeader } from "../components/SettingsPageheader";
 import { SelectMenuBasic } from "../components/SelectMenuBasic";
-
-import { SettingsItem } from "./devices.$id.settings";
 
 export default function SettingsAppearanceRoute() {
   const [currentTheme, setCurrentTheme] = useState(() => {
