+++ ui/src/routes/devices.$id.settings.general._index.tsx
@@ -1,6 +1,7 @@
 
 import { useState , useEffect } from "react";
 
+import { SettingsItem } from "@components/SettingsItem";
 import { JsonRpcResponse, useJsonRpc } from "@/hooks/useJsonRpc";
 
 import { SettingsPageHeader } from "../components/SettingsPageheader";
@@ -10,7 +11,6 @@ import Checkbox from "../components/Checkbox";
 import { useDeviceUiNavigation } from "../hooks/useAppNavigation";
 import { useDeviceStore } from "../hooks/stores";
 
-import { SettingsItem } from "./devices.$id.settings";
 
 export default function SettingsGeneralRoute() {
   const { send } = useJsonRpc();
