+++ ui/src/routes/devices.$id.settings.video.tsx
@@ -3,13 +3,12 @@ import { useCallback, useEffect, useState } from "react";
 import { Button } from "@/components/Button";
 import { TextAreaWithLabel } from "@/components/TextArea";
 import { JsonRpcResponse, useJsonRpc } from "@/hooks/useJsonRpc";
+import { SettingsItem } from "@components/SettingsItem";
 import { SettingsPageHeader } from "@components/SettingsPageheader";
 import { useSettingsStore } from "@/hooks/stores";
 import { SelectMenuBasic } from "@components/SelectMenuBasic";
 import Fieldset from "@components/Fieldset";
 import notifications from "@/notifications";
-
-import { SettingsItem } from "./devices.$id.settings";
 
 const defaultEdid =
   "00ffffffffffff0052620188008888881c150103800000780a0dc9a05747982712484c00000001010101010101010101010101010101023a801871382d40582c4500c48e2100001e011d007251d01e206e285500c48e2100001e000000fc00543734392d6648443732300a20000000fd00147801ff1d000a202020202020017b";
